requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-js-asset==2.0.0
django-mptt==0.14.0
factory-boy==3.2.1
Faker==15.1.3
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
